Move a few other common helpers in the common lib
